Create Robot Arm
It has been always my childhood dream to make a robotic arm and this semester EEE 102 course empowered me with the necessary knowledge to build a robotic arm from scratch. In these times of quarantine, I wasn't able to go to shops and look for material to build a robotic arm so I recycled cardboard pieces and used cheap servo motor that I already had. It was hard but a great learning experience