Add files via upload
This course project for digital logic design 102 that I made is a robotic arm named “Axis”. The robotic arm has 3 degrees of freedom. It consists of three servomotors attached at each joint and one servo attached to operate claw. This project is basically a prototype of industrial multipurpose use robotic arm with lightweight works so it has an IR obstacle sensor attached to it in order to make sure that if somebody comes close to its proximity it stops working. The 4 x4 keypad matrix is used as a control Panel.